Fixed `scroll` compatability with numpy 1.12

+++ moviepy/video/fx/scroll.py
@@ -1,5 +1,3 @@
-import numpy as np
-
 def scroll(clip, h=None, w=None, x_speed=0, y_speed=0,
            x_start=0, y_start=0, apply_to="mask",
            ease_func=None, period=1):
@@ -14,8 +12,8 @@ def scroll(clip, h=None, w=None, x_speed=0, y_speed=0,
     def f(gf,t):
         if ease_func and t < period:
             t = ease_func(t, duration=period) * period
-        x = max(0, min(xmax, x_start+ round(x_speed*t)))
-        y = max(0, min(ymax, y_start+ round(y_speed*t)))
+        x = int(max(0, min(xmax, x_start+ round(x_speed*t))))
+        y = int(max(0, min(ymax, y_start+ round(y_speed*t))))
         return gf(t)[y:y+h, x:x+w]
     
     return clip.fl(f, apply_to = apply_to)
